+++ qualit-allure/src/main/java/ru/ibsqa/qualit/AllureJUnit.java
@@ -5,13 +5,21 @@ import org.junit.platform.engine.TestExecutionResult;
 import org.junit.platform.engine.reporting.ReportEntry;
 import org.junit.platform.launcher.TestIdentifier;
 import org.junit.platform.launcher.TestPlan;
+import org.springframework.core.io.Resource;
+import ru.ibsqa.qualit.utils.spring.SpringUtils;
+
+import java.util.Optional;
 
 public class AllureJUnit extends AllureJunitPlatform {
     private final boolean allureEnabled = Boolean.parseBoolean(System.getProperty("allure.enabled", "true"));
+    private final Resource[] resources = SpringUtils.getResources("classpath:/cucumber.properties");
+    private final boolean cucumberEnabled = Optional.ofNullable(resources)
+            .map(resource -> resource.length > 0 && resource[0].exists())
+            .orElse(false);
 
     @Override
     public void testPlanExecutionStarted(TestPlan testPlan) {
-        if (!allureEnabled) {
+        if (!allureEnabled || cucumberEnabled) {
             return;
         }
 
@@ -20,7 +28,7 @@ public class AllureJUnit extends AllureJunitPlatform {
 
     @Override
     public void testPlanExecutionFinished(TestPlan testPlan) {
-        if (!allureEnabled) {
+        if (!allureEnabled || cucumberEnabled) {
             return;
         }
 
@@ -29,7 +37,7 @@ public class AllureJUnit extends AllureJunitPlatform {
 
     @Override
     public void executionStarted(TestIdentifier testIdentifier) {
-        if (!allureEnabled) {
+        if (!allureEnabled || cucumberEnabled) {
             return;
         }
 
@@ -38,7 +46,7 @@ public class AllureJUnit extends AllureJunitPlatform {
 
     @Override
     public void executionFinished(TestIdentifier testIdentifier, TestExecutionResult testExecutionResult) {
-        if (!allureEnabled) {
+        if (!allureEnabled || cucumberEnabled) {
             return;
         }
 
@@ -47,7 +55,7 @@ public class AllureJUnit extends AllureJunitPlatform {
 
     @Override
     public void executionSkipped(TestIdentifier testIdentifier, String reason) {
-        if (!allureEnabled) {
+        if (!allureEnabled || cucumberEnabled) {
             return;
         }
 
@@ -56,7 +64,7 @@ public class AllureJUnit extends AllureJunitPlatform {
 
     @Override
     public void reportingEntryPublished(TestIdentifier testIdentifier, ReportEntry entry) {
-        if (!allureEnabled) {
+        if (!allureEnabled || cucumberEnabled) {
             return;
         }